backend/requirements.txt
flask==3.0.3
flask-cors==4.0.1
numpy==1.26.4
pillow==10.3.0
opencv-python==4.9.0.80
matplotlib==3.8.4
scikit-learn==1.5.0
gunicorn==22.0.0
torch
torchvision
transformers
